packages/mcp-servers/google-drive/pyproject.toml
[project]
name = "neuralflow-google-drive-mcp"
version = "0.1.0"
description = "Google Drive MCP server for NeuralFlow."
requires-python = ">=3.12"
dependencies = [
    "mcp",
    "google-api-python-client",
    "google-auth-oauthlib",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["server"]
